client: update examples
- Move "Example" tests separate so that we can use the actual import
- Add "WithUserAgent" in the examples, which should be best practice
