+++ client/service_logs_test.go
@@ -1,15 +1,10 @@
 package client
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
-	"os"
 	"testing"
-	"time"
 
 	cerrdefs "github.com/containerd/errdefs"
 	"gotest.tools/v3/assert"
@@ -139,25 +134,5 @@ func TestServiceLogs(t *testing.T) {
 			assert.NilError(t, err)
 			assert.Check(t, is.Contains(string(content), "response"))
 		})
-	}
-}
-
-func ExampleClient_ServiceLogs_withTimeout() {
-	client, err := New(FromEnv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	res, err := client.ServiceLogs(ctx, "service_id", ServiceLogsOptions{})
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer res.Close()
-
-	_, err = io.Copy(os.Stdout, res)
-	if err != nil && !errors.Is(err, io.EOF) {
-		log.Fatal(err)
 	}
 }
